merge
Add IssuanceItems functionality and related updates

- Introduced new permissions and resource for IssuanceItems.
- Updated database migrations to include IssuanceItem table.
- Modified commands and handlers for IssuanceItem creation and validation.
- Enhanced Inventory management and updated related classes.
- Implemented endpoints for IssuanceItem CRUD operations.
- Adjusted product structure and seeding logic for new properties.

+++ api/modules/Catalog/Catalog.Application/PurchaseItems/Create/v1/CreatePurchaseItemCommand.cs
@@ -5,6 +5,6 @@ namespace AMIS.WebApi.Catalog.Application.PurchaseItems.Create.v1;
 public sealed record CreatePurchaseItemCommand(
     Guid PurchaseId,
     Guid ProductId,
-    [property: DefaultValue(1)] decimal Qty,
+    [property: DefaultValue(1)] int Qty,
     [property: DefaultValue(0)] decimal UnitPrice,
     [property: DefaultValue("Pending")] string? Status = "Pending") : IRequest<CreatePurchaseItemResponse>;
