+++ api/modules/Catalog/Catalog.Application/PurchaseItems/Get/v1/PurchaseItemResponse.cs
@@ -1,4 +1,4 @@
 using AMIS.WebApi.Catalog.Application.Products.Get.v1;
 
 namespace AMIS.WebApi.Catalog.Application.PurchaseItems.Get.v1;
-public sealed record PurchaseItemResponse(Guid? Id, Guid PurchaseId, Guid ProductId, decimal Qty, decimal UnitPrice, string Status, ProductResponse? Product);
+public sealed record PurchaseItemResponse(Guid? Id, Guid PurchaseId, Guid ProductId, int Qty, decimal UnitPrice, string Status, ProductResponse? Product);
